requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyHamcrest==2.0.3
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.8
